check explicitly for file = null
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@10237 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/io/FileBucket.java
@@ -45,6 +45,8 @@ public class FileBucket implements Bucket, SerializableToFieldSetBucket {
 		this.file = file;
 		this.deleteOnFinalize = deleteOnFinalize;
 		this.deleteOnFree = deleteOnFree;
+		if(file == null)
+			throw new NullPointerException();
 		if(deleteOnExit)
 			file.deleteOnExit();
 		// Useful for finding temp file leaks.
